+++ apps/router-demo/router-remote4-2004/src/bootstrap.tsx
@@ -1,9 +1,9 @@
 import React from 'react';
-import ReactDOM from 'react-dom/client';
+import { createRoot } from 'react-dom/client';
 import App from './App';
 
 //@ts-ignore
-const root = ReactDOM.createRoot(document.getElementById('root')!);
+const root = createRoot(document.getElementById('root')!);
 root.render(
   <React.StrictMode>
     <App />