+++ packages/cli/src/utils/readConfig.ts
@@ -1,6 +1,7 @@
 import path from 'path';
 import { bundle } from '@modern-js/node-bundle-require';
 import type { moduleFederationPlugin } from '@module-federation/sdk';
+import { pathToFileURL } from 'url';
 
 const DEFAULT_CONFIG_PATH = 'module-federation.config.ts';
 
@@ -15,7 +16,7 @@ export const getConfigPath = (userConfigPath?: string) => {
 export async function readConfig(userConfigPath?: string) {
   const configPath = getConfigPath(userConfigPath);
   const preBundlePath = await bundle(configPath);
-  const mfConfig = (await import(preBundlePath)).default
+  const mfConfig = (await import(pathToFileURL(preBundlePath).href)).default
     .default as unknown as moduleFederationPlugin.ModuleFederationPluginOptions;
   return mfConfig;
 }